requirements.txt
#scrapy
#pymongo
#requests
python-dotenv
#parse_rest
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
enum34==1.1.6
functools32==3.2.3.post2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.23
lxml==4.4.2
#parse-rest==0.2.20170114
parse-rest==0.2.20141004
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.10.1
pymongo==3.10.1
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.14.0
Twisted==19.10.0
urllib3==1.25.8
virtualenv==16.7.9
w3lib==1.21.0
zope.interface==4.7.1